backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
pyotp==2.9.0
cryptography==41.0.7
gunicorn==21.2.0
pytest==7.4.3
pytest-mock==3.12.0

# ML Libraries
torch==2.1.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
tensorflow==2.15.0

# WebSocket Support
flask-socketio==5.3.5
python-socketio==5.10.0
eventlet==0.33.3

# Redis for Caching
redis==5.0.1
flask-caching==2.1.0

# Message Queue (Celery + RabbitMQ)
celery==5.3.4
kombu==5.3.4

# Blockchain Integration
web3==6.11.3
py-solc-x==2.0.2

# IPFS Integration
ipfshttpclient==0.8.0a2

# AI Integration (Claude API)
anthropic==0.7.8

# Geolocation and IP Reputation
geoip2==4.7.0
requests==2.31.0

# NLP Libraries
nltk==3.8.1
spacy==3.7.2

# Additional Utilities
geopy==2.4.1
python-dateutil==2.8.2
